Add Home and search pages

+++ src/main/resources/db/migration/V1__BaseMigrationForCreateTable.sql
@@ -8,17 +8,17 @@ CREATE TABLE IF NOT EXISTS Users
 CREATE TABLE IF NOT EXISTS Locations
 (
     id        SERIAL PRIMARY KEY,
-    name      VARCHAR(255)   NOT NULL,
-    user_id   INTEGER        NOT NULL,
-    latitude  DECIMAL(10, 8) NOT NULL,
-    longitude DECIMAL(10, 8) NOT NULL,
+    name      VARCHAR(255)  NOT NULL,
+    user_id   INTEGER       NOT NULL,
+    latitude  DECIMAL NOT NULL,
+    longitude DECIMAL NOT NULL,
     FOREIGN KEY (user_id) REFERENCES Users (id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS Sessions
 (
     id        uuid PRIMARY KEY,
-    user_id   INTEGER   NOT NULL,
+    user_id   INTEGER                  NOT NULL,
     expiresAt TIMESTAMP WITH TIME ZONE NOT NULL,
     FOREIGN KEY (User_id) REFERENCES Users (id) ON DELETE CASCADE
 );
